refactor meetup to data
