+++ data/meetup.json
@@ -1,0 +1,11 @@
+{
+  "emails": {
+    "contact": "vuejsfa@gmail.com",
+    "sponsors": "vuejsfa@gmail.com",
+    "article": "vuejsfa@gmail.com"
+  },
+  "social": {
+    "twitter": "https://twitter.com/vuejs_fa",
+    "telegram": "https://t.me/vuejs_fa"
+  }
+}
