+++ pages/index.vue
@@ -245,9 +245,10 @@
 </template>
 
 <script>
-import meetup from '../config/meetup'
+import meetup from '../data/meetup'
 import speakers from '../data/speakers'
 import sponsors from '../data/sponsors'
+
 // import Badge from '../components/global/badge'
 import Brand from '../components/custom/index/brand'
